Stepping through the program and debuggin'
Tried running it and noticed that it wasn't even able to run the first
instruction, so I went ahead and started debugging from there.

Added lots of debug messages (switched on/off with the #define at the top). The
messages are useful when debugging a specific function, but they could probably
be more informative (as it is, it looks like they just give you random bits of
information).

The first couple of instructions appear to work. addi definitely works
perfectly. Something is wrong with Jump, but we can look into that more later.