Cleaning up project.c before submitting
Needed to disable the debug prints. Also cleared out some of the remianing TODO
blocks. Tested it once more and it looks good to me.

Also needed to put the MEMSIZE define in project.c since we weren't supposed to
modify spimcore.c or spimcore.h according to the instructions.